Refactored State demo
