--- components/doc/datatable/state/localdoc.js
@@ -1,436 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import { CustomerService } from '../../../../service/CustomerService';
-import { FilterMatchMode, FilterOperator } from '../../../lib/api/Api';
-import { Column } from '../../../lib/column/Column';
-import { DataTable } from '../../../lib/datatable/DataTable';
-import { Dropdown } from '../../../lib/dropdown/Dropdown';
-import { InputText } from '../../../lib/inputtext/InputText';
-import { MultiSelect } from '../../../lib/multiselect/MultiSelect';
-import { DocSectionCode } from '../../common/docsectioncode';
-import { DocSectionText } from '../../common/docsectiontext';
-
-export function LocalStorageDoc(props) {
-    const [customers, setCustomers] = useState(null);
-    const [filters, setFilters] = useState({
-        global: { value: null, matchMode: FilterMatchMode.CONTAINS },
-        name: { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.STARTS_WITH }] },
-        'country.name': { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.STARTS_WITH }] },
-        representative: { value: null, matchMode: FilterMatchMode.IN },
-        status: { operator: FilterOperator.OR, constraints: [{ value: null, matchMode: FilterMatchMode.EQUALS }] }
-    });
-
-    const [selectedCustomer, setSelectedCustomer] = useState(null);
-    const representatives = [
-        { name: 'Amy Elsner', image: 'amyelsner.png' },
-        { name: 'Anna Fali', image: 'annafali.png' },
-        { name: 'Asiya Javayant', image: 'asiyajavayant.png' },
-        { name: 'Bernardo Dominic', image: 'bernardodominic.png' },
-        { name: 'Elwin Sharvill', image: 'elwinsharvill.png' },
-        { name: 'Ioni Bowcher', image: 'ionibowcher.png' },
-        { name: 'Ivan Magalhaes', image: 'ivanmagalhaes.png' },
-        { name: 'Onyama Limba', image: 'onyamalimba.png' },
-        { name: 'Stephen Shaw', image: 'stephenshaw.png' },
-        { name: 'XuXue Feng', image: 'xuxuefeng.png' }
-    ];
-    const statuses = ['unqualified', 'qualified', 'new', 'negotiation', 'renewal', 'proposal'];
-
-    useEffect(() => {
-        CustomerService.getCustomersMedium().then((data) => setCustomers(data));
-    }, []); // eslint-disable-line react-hooks/exhaustive-deps
-
-    const countryBodyTemplate = (rowData) => {
-        return (
-            <React.Fragment>
-                <img alt={rowData.country.code} src="https://primefaces.org/cdn/primereact/images/flag/flag_placeholder.png" className={`flag flag-${rowData.country.code}`} width="30" />
-                <span className="vertical-align-middle ml-2">{rowData.country.name}</span>
-            </React.Fragment>
-        );
-    };
-
-    const representativeBodyTemplate = (rowData) => {
-        return (
-            <React.Fragment>
-                <img alt={rowData.representative.name} src={`https://primefaces.org/cdn/primereact/images/avatar/${rowData.representative.image}`} width="32" style={{ verticalAlign: 'middle' }} />
-                <span className="vertical-align-middle ml-2">{rowData.representative.name}</span>
-            </React.Fragment>
-        );
-    };
-
-    const representativeFilterTemplate = (options) => {
-        return <MultiSelect value={options.value} options={representatives} itemTemplate={representativesItemTemplate} onChange={(e) => options.filterCallback(e.value)} optionLabel="name" placeholder="Any" className="p-column-filter" />;
-    };
-
-    const representativesItemTemplate = (option) => {
-        return (
-            <div className="p-multiselect-representative-option">
-                <img alt={option.name} src={`https://primefaces.org/cdn/primereact/images/avatar/${option.image}`} width={32} style={{ verticalAlign: 'middle' }} />
-                <span className="vertical-align-middle ml-2">{option.name}</span>
-            </div>
-        );
-    };
-
-    const statusBodyTemplate = (rowData) => {
-        return <span className={`customer-badge status-${rowData.status}`}>{rowData.status}</span>;
-    };
-
-    const statusFilterTemplate = (options) => {
-        return <Dropdown value={options.value} options={statuses} onChange={(e) => options.filterCallback(e.value, options.index)} itemTemplate={statusItemTemplate} placeholder="Select a Status" className="p-column-filter" showClear />;
-    };
-
-    const statusItemTemplate = (option) => {
-        return <span className={`customer-badge status-${option}`}>{option}</span>;
-    };
-
-    const onGlobalFilterChange = (event) => {
-        const value = event.target.value;
-        let _filters = { ...filters };
-
-        _filters['global'].value = value;
-
-        setFilters(_filters);
-    };
-
-    const renderHeader = () => {
-        const value = filters['global'] ? filters['global'].value : '';
-
-        return (
-            <span className="p-input-icon-left">
-                <i className="pi pi-search" />
-                <InputText type="search" value={value || ''} onChange={(e) => onGlobalFilterChange(e)} placeholder="Global Search" />
-            </span>
-        );
-    };
-
-    const header = renderHeader();
-
-    const code = {
-        basic: `
-<DataTable value={customers} paginator rows={10} header={header} filters={filters} onFilter={(e) => setFilters(e.filters)}
-    selection={selectedCustomer} onSelectionChange={e => setSelectedCustomer(e.value)} selectionMode="single" dataKey="id" responsiveLayout="scroll"
-    stateStorage="local" stateKey="dt-state-demo-local" emptyMessage="No customers found.">
-    <Column field="name" header="Name" sortable filter filterPlaceholder="Search by name"></Column>
-    <Column header="Country" body={countryBodyTemplate} sortable sortField="country.name" filter filterField="country.name" filterPlaceholder="Search by country"></Column>
-    <Column header="Agent" body={representativeBodyTemplate} sortable sortField="representative.name" filter filterField="representative" showFilterMatchModes={false} filterElement={representativeFilterTemplate} filterMenuStyle={{ width: '14rem' }}></Column>
-    <Column field="status" header="Status" body={statusBodyTemplate} sortable filter filterElement={statusFilterTemplate} filterMenuStyle={{ width: '14rem' }}></Column>
-</DataTable>
-        `,
-        javascript: `
-import React, { useState, useEffect } from 'react';
-import { FilterMatchMode, FilterOperator } from 'primereact/api';
-import { DataTable } from 'primereact/datatable';
-import { Column } from 'primereact/column';
-import { InputText } from 'primereact/inputtext';
-import { MultiSelect } from 'primereact/multiselect';
-import { Dropdown } from 'primereact/dropdown';
-import { CustomerService } from './service/CustomerService';
-
-const LocalStorageDoc = () => {
-
-    const [customers, setCustomers] = useState(null);
-    const [filters, setFilters] = useState({
-        global: { value: null, matchMode: FilterMatchMode.CONTAINS },
-        name: { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.STARTS_WITH }] },
-        'country.name': { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.STARTS_WITH }] },
-        representative: { value: null, matchMode: FilterMatchMode.IN },
-        status: { operator: FilterOperator.OR, constraints: [{ value: null, matchMode: FilterMatchMode.EQUALS }] }
-    });
-
-    const [selectedCustomer, setSelectedCustomer] = useState(null);
-    const representatives = [
-        { name: 'Amy Elsner', image: 'amyelsner.png' },
-        { name: 'Anna Fali', image: 'annafali.png' },
-        { name: 'Asiya Javayant', image: 'asiyajavayant.png' },
-        { name: 'Bernardo Dominic', image: 'bernardodominic.png' },
-        { name: 'Elwin Sharvill', image: 'elwinsharvill.png' },
-        { name: 'Ioni Bowcher', image: 'ionibowcher.png' },
-        { name: 'Ivan Magalhaes', image: 'ivanmagalhaes.png' },
-        { name: 'Onyama Limba', image: 'onyamalimba.png' },
-        { name: 'Stephen Shaw', image: 'stephenshaw.png' },
-        { name: 'XuXue Feng', image: 'xuxuefeng.png' }
-    ];
-    const statuses = ['unqualified', 'qualified', 'new', 'negotiation', 'renewal', 'proposal'];
-
-    useEffect(() => {
-        CustomerService.getCustomersMedium().then((data) => setCustomers(data));
-    }, []); // eslint-disable-line react-hooks/exhaustive-deps
-
-    const countryBodyTemplate = (rowData) => {
-        return (
-            <React.Fragment>
-                <img alt={rowData.country.code} src="https://primefaces.org/cdn/primereact/images/flag/flag_placeholder.png" className={\`flag flag-\${rowData.country.code}\`} width="30" />
-                <span className="vertical-align-middle ml-2">{rowData.country.name}</span>
-            </React.Fragment>
-        );
-    }
-
-    const representativeBodyTemplate = (rowData) => {
-        return (
-            <React.Fragment>
-                <img alt={rowData.representative.name} src={\`https://primefaces.org/cdn/primereact/images/avatar/\${rowData.representative.image}\`} width="32" style={{ verticalAlign: 'middle' }} />
-                <span className="vertical-align-middle ml-2">{rowData.representative.name}</span>
-            </React.Fragment>
-        );
-    }
-
-    const representativeFilterTemplate = (options) => {
-        return <MultiSelect value={options.value} options={representatives} itemTemplate={representativesItemTemplate} onChange={(e) => options.filterCallback(e.value)} optionLabel="name" placeholder="Any" className="p-column-filter" />;
-    };
-
-    const representativesItemTemplate = (option) => {
-        return (
-            <div className="p-multiselect-representative-option">
-                <img alt={option.name} src={\`https://primefaces.org/cdn/primereact/images/avatar/\${option.image}\`} width={32} style={{ verticalAlign: 'middle' }} />
-                <span className="vertical-align-middle ml-2">{option.name}</span>
-            </div>
-        );
-    }
-
-    const statusBodyTemplate = (rowData) => {
-        return <span className={\`customer-badge status-\${rowData.status}\`}>{rowData.status}</span>;
-    }
-
-    const statusFilterTemplate = (options) => {
-        return <Dropdown value={options.value} options={statuses} onChange={(e) => options.filterCallback(e.value, options.index)} itemTemplate={statusItemTemplate} placeholder="Select a Status" className="p-column-filter" showClear />;
-    }
-
-    const statusItemTemplate = (option) => {
-        return <span className={\`customer-badge status-\${option}\`}>{option}</span>;
-    }
-
-    const onGlobalFilterChange = (event) => {
-        const value = event.target.value;
-        let _filters = { ...filters };
-
-        _filters['global'].value = value;
-
-        setFilters(_filters);
-    };
-
-    const renderHeader = () => {
-        const value = filters['global'] ? filters['global'].value : '';
-
-        return (
-            <span className="p-input-icon-left">
-                <i className="pi pi-search" />
-                <InputText type="search" value={value || ''} onChange={(e) => onGlobalFilterChange(e)} placeholder="Global Search" />
-            </span>
-        );
-    };
-
-    const header = renderHeader();
-
-    return (
-        <div className="card">
-            <DataTable value={customers} paginator rows={10} header={header} filters={filters} onFilter={(e) => setFilters(e.filters)}
-                selection={selectedCustomer} onSelectionChange={e => setSelectedCustomer(e.value)} selectionMode="single" dataKey="id" responsiveLayout="scroll"
-                stateStorage="local" stateKey="dt-state-demo-local" emptyMessage="No customers found.">
-                <Column field="name" header="Name" sortable filter filterPlaceholder="Search by name"></Column>
-                <Column header="Country" body={countryBodyTemplate} sortable sortField="country.name" filter filterField="country.name" filterPlaceholder="Search by country"></Column>
-                <Column header="Agent" body={representativeBodyTemplate} sortable sortField="representative.name" filter filterField="representative" showFilterMatchModes={false} filterElement={representativeFilterTemplate} filterMenuStyle={{ width: '14rem' }}></Column>
-                <Column field="status" header="Status" body={statusBodyTemplate} sortable filter filterElement={statusFilterTemplate} filterMenuStyle={{ width: '14rem' }}></Column>
-            </DataTable>
-        </div>
-    );
-}
-        `,
-        typescript: `
-import React, { useState, useEffect } from 'react';
-import { FilterMatchMode, FilterOperator } from 'primereact/api';
-import { DataTable } from 'primereact/datatable';
-import { Column } from 'primereact/column';
-import { InputText } from 'primereact/inputtext';
-import { MultiSelect } from 'primereact/multiselect';
-import { Dropdown } from 'primereact/dropdown';
-import { CustomerService } from './service/CustomerService';
-
-const LocalStorageDoc = () => {
-
-    const [customers, setCustomers] = useState(null);
-    const [filters, setFilters] = useState({
-        global: { value: null, matchMode: FilterMatchMode.CONTAINS },
-        name: { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.STARTS_WITH }] },
-        'country.name': { operator: FilterOperator.AND, constraints: [{ value: null, matchMode: FilterMatchMode.STARTS_WITH }] },
-        representative: { value: null, matchMode: FilterMatchMode.IN },
-        status: { operator: FilterOperator.OR, constraints: [{ value: null, matchMode: FilterMatchMode.EQUALS }] }
-    });
-    const [selectedCustomer, setSelectedCustomer] = useState(null);
-    const representatives = [
-        { name: "Amy Elsner", image: 'amyelsner.png' },
-        { name: "Anna Fali", image: 'annafali.png' },
-        { name: "Asiya Javayant", image: 'asiyajavayant.png' },
-        { name: "Bernardo Dominic", image: 'bernardodominic.png' },
-        { name: "Elwin Sharvill", image: 'elwinsharvill.png' },
-        { name: "Ioni Bowcher", image: 'ionibowcher.png' },
-        { name: "Ivan Magalhaes", image: 'ivanmagalhaes.png' },
-        { name: "Onyama Limba", image: 'onyamalimba.png' },
-        { name: "Stephen Shaw", image: 'stephenshaw.png' },
-        { name: "XuXue Feng", image: 'xuxuefeng.png' }
-    ];
-    const statuses = [
-        'unqualified', 'qualified', 'new', 'negotiation', 'renewal', 'proposal'
-    ];
-    
-
-    const filtersMap = {
-        'filters': { value: filters, callback: setFilters },
-        'filters': { value: filters, callback: setFilters },
-        'filters3': { value: filters3, callback: setFilters3 },
-    };
-
-    useEffect(() => {
-        CustomerService.getCustomersMedium().then(data => setCustomers(data));
-    }, []); // eslint-disable-line react-hooks/exhaustive-deps
-
-    const onCustomSaveState = (state) => {
-        sessionStorage.setItem('dt-state-demo-custom', JSON.stringify(state));
-    }
-
-    const onCustomRestoreState = () => {
-        return JSON.parse(sessionStorage.getItem('dt-state-demo-custom'));
-    }
-
-    const countryBodyTemplate = (rowData) => {
-        return (
-            <React.Fragment>
-                <img alt={rowData.country.code} src="https://primefaces.org/cdn/primereact/images/flag/flag_placeholder.png" className={\`flag flag-\${rowData.country.code}\`} width="30" />
-                <span className="vertical-align-middle ml-2">{rowData.country.name}</span>
-            </React.Fragment>
-        );
-    }
-
-    const representativeBodyTemplate = (rowData) => {
-        return (
-            <React.Fragment>
-                <img alt={rowData.representative.name} src={\`https://primefaces.org/cdn/primereact/images/avatar/\${rowData.representative.image}\`} width="32" style={{ verticalAlign: 'middle' }} />
-                <span className="vertical-align-middle ml-2">{rowData.representative.name}</span>
-            </React.Fragment>
-        );
-    }
-
-    const representativeFilterTemplate = (options) => {
-        return <MultiSelect value={options.value} options={representatives} itemTemplate={representativesItemTemplate} onChange={(e) => options.filterCallback(e.value)} optionLabel="name" placeholder="Any" className="p-column-filter" />;
-    };
-
-    const representativesItemTemplate = (option) => {
-        return (
-            <div className="p-multiselect-representative-option">
-                <img alt={option.name} src={\`https://primefaces.org/cdn/primereact/images/avatar/\${option.image}\`} width={32} style={{ verticalAlign: 'middle' }} />
-                <span className="vertical-align-middle ml-2">{option.name}</span>
-            </div>
-        );
-    }
-
-    const statusBodyTemplate = (rowData) => {
-        return <span className={\`customer-badge status-\${rowData.status}\`}>{rowData.status}</span>;
-    }
-
-    const statusFilterTemplate = (options) => {
-        return <Dropdown value={options.value} options={statuses} onChange={(e) => options.filterCallback(e.value, options.index)} itemTemplate={statusItemTemplate} placeholder="Select a Status" className="p-column-filter" showClear />;
-    }
-
-    const statusItemTemplate = (option) => {
-        return <span className={\`customer-badge status-\${option}\`}>{option}</span>;
-    }
-
-    const onGlobalFilterChange = (event) => {
-        const value = event.target.value;
-        let _filters = { ...filters };
-
-        _filters['global'].value = value;
-
-        setFilters(_filters);
-    };
-
-    const renderHeader = () => {
-        const value = filters['global'] ? filters['global'].value : '';
-
-        return (
-            <span className="p-input-icon-left">
-                <i className="pi pi-search" />
-                <InputText type="search" value={value || ''} onChange={(e) => onGlobalFilterChange(e)} placeholder="Global Search" />
-            </span>
-        );
-    }
-
-    const header = renderHeader();
-
-    return (
-        <div className="card">
-            <DataTable value={customers} paginator rows={10} header={header} filters={filters} onFilter={(e) => setFilters(e.filters)}
-                selection={selectedCustomer} onSelectionChange={e => setSelectedCustomer(e.value)} selectionMode="single" dataKey="id" responsiveLayout="scroll"
-                stateStorage="local" stateKey="dt-state-demo-local" emptyMessage="No customers found.">
-                <Column field="name" header="Name" sortable filter filterPlaceholder="Search by name"></Column>
-                <Column header="Country" body={countryBodyTemplate} sortable sortField="country.name" filter filterField="country.name" filterPlaceholder="Search by country"></Column>
-                <Column header="Agent" body={representativeBodyTemplate} sortable sortField="representative.name" filter filterField="representative" showFilterMatchModes={false} filterElement={representativeFilterTemplate} filterMenuStyle={{ width: '14rem' }}></Column>
-                <Column field="status" header="Status" body={statusBodyTemplate} sortable filter filterElement={statusFilterTemplate} filterMenuStyle={{ width: '14rem' }}></Column>
-            </DataTable>
-        </div>
-    );
-}
-        `,
-        data: `
-/* CustomerService */ 
-{
-    id: 1000,
-    name: 'James Butt',
-    country: {
-        name: 'Algeria',
-        code: 'dz'
-    },
-    company: 'Benton, John B Jr',
-    date: '2015-09-13',
-    status: 'unqualified',
-    verified: true,
-    activity: 17,
-    representative: {
-        name: 'Ioni Bowcher',
-        image: 'ionibowcher.png'
-    },
-    balance: 70663
-},
-...
-       `
-    };
-
-    return (
-        <>
-            <DocSectionText {...props}>
-                <p>Local Storage.</p>
-            </DocSectionText>
-            <div className="card">
-                <DataTable
-                    value={customers}
-                    paginator
-                    rows={10}
-                    header={header}
-                    filters={filters}
-                    onFilter={(e) => setFilters(e.filters)}
-                    selection={selectedCustomer}
-                    onSelectionChange={(e) => setSelectedCustomer(e.value)}
-                    selectionMode="single"
-                    dataKey="id"
-                    responsiveLayout="scroll"
-                    stateStorage="local"
-                    stateKey="dt-state-demo-local"
-                    emptyMessage="No customers found."
-                >
-                    <Column field="name" header="Name" sortable filter filterPlaceholder="Search by name"></Column>
-                    <Column header="Country" body={countryBodyTemplate} sortable sortField="country.name" filter filterField="country.name" filterPlaceholder="Search by country"></Column>
-                    <Column
-                        header="Agent"
-                        body={representativeBodyTemplate}
-                        sortable
-                        sortField="representative.name"
-                        filter
-                        filterField="representative"
-                        showFilterMatchModes={false}
-                        filterElement={representativeFilterTemplate}
-                        filterMenuStyle={{ width: '14rem' }}
-                    ></Column>
-                    <Column field="status" header="Status" body={statusBodyTemplate} sortable filter filterElement={statusFilterTemplate} filterMenuStyle={{ width: '14rem' }}></Column>
-                </DataTable>
-            </div>
-            <DocSectionCode code={code} service={['CustomerService']} />
-        </>
-    );
-}
